Update existing services
